Treat elapsed time as signed, since rb_player_get_time returns a signed
2006-10-28  Jonathan Matthew  <jonathan@kaolin.wh9.net>

	* shell/rb-shell-player.c: (rb_shell_player_class_init),
	(rb_shell_player_sync_with_source),
	(rb_shell_player_get_playing_time_string), (rb_shell_player_seek),
	(tick_cb):
	* widgets/rb-header.c: (rb_header_sync_time),
	(rb_header_elapsed_changed_cb):
	Treat elapsed time as signed, since rb_player_get_time returns a
	signed value.  When actually using the value, mostly treat negative
	values as 0.  Fixes #365908.

+++ widgets/rb-header.c
@@ -57,7 +57,7 @@ static gboolean slider_moved_callback (GtkWidget *widget, GdkEventMotion *event,
 static gboolean slider_release_callback (GtkWidget *widget, GdkEventButton *event, RBHeader *header);
 static void slider_changed_callback (GtkWidget *widget, RBHeader *header);
 
-static void rb_header_elapsed_changed_cb (RBShellPlayer *player, guint elapsed, RBHeader *header);
+static void rb_header_elapsed_changed_cb (RBShellPlayer *player, long elapsed, RBHeader *header);
 
 struct RBHeaderPrivate
 {
@@ -471,7 +471,7 @@ rb_header_set_show_timeline (RBHeader *header,
 gboolean
 rb_header_sync_time (RBHeader *header)
 {
-	int seconds;
+	long seconds;
 
 	if (header->priv->shell_player == NULL)
 		return TRUE;
@@ -487,7 +487,7 @@ rb_header_sync_time (RBHeader *header)
 		double progress = 0.0;
 
 		if (seconds > 0) {
-			progress = (double) (long) seconds;
+			progress = (double) seconds;
 		} else {
 			header->priv->adjustment->upper = header->priv->duration;
 			g_signal_emit_by_name (G_OBJECT (header->priv->adjustment), "changed");
@@ -655,7 +655,7 @@ rb_header_update_elapsed (RBHeader *header)
 
 static void
 rb_header_elapsed_changed_cb (RBShellPlayer *player,
-			      guint elapsed,
+			      long elapsed,
 			      RBHeader *header)
 {
 	header->priv->elapsed_time = elapsed;
